Remove cython from source code and tests
