+++ setup.py
@@ -20,23 +20,12 @@
 
 import os
 
-from setuptools import Extension, find_packages, setup
+from setuptools import find_packages, setup
 
 with open("README.rst", "r") as readme_file:
     LONG_DESCRIPTION = readme_file.read()
 
 NO_CYTHON = bool(os.environ.get("NO_CYTHON"))
-
-
-def modules():
-    if NO_CYTHON:
-        return []
-    else:
-        return [
-            Extension("fastqsplitter.split_cy",
-                      ["src/fastqsplitter/split_cy.pyx"])
-        ]
-
 
 setup(
     name="fastqsplitter",
@@ -52,7 +41,6 @@ setup(
     packages=find_packages('src'),
     package_dir={'': 'src'},
     # This makes sure source is included in the binary distribution
-    package_data={'fastqsplitter': ['*.pyx']},
     url="https://github.com/LUMC/fastqsplitter",
     classifiers=[
         "Programming Language :: Python :: 3 :: Only",
@@ -65,7 +53,6 @@ setup(
         "Topic :: Scientific/Engineering :: Bio-Informatics"
     ],
     python_requires=">=3.5",  # Because we use type annotation.
-    setup_requires=[] if NO_CYTHON else ["cython"],
     install_requires=[
        "xopen>=0.8.1"
     ],
@@ -73,6 +60,5 @@ setup(
         "console_scripts": [
             'fastqsplitter=fastqsplitter:main'
         ]
-    },
-    ext_modules=modules()
+    }
 )
