+++ tox.ini
@@ -7,8 +7,6 @@ deps=coverage
      pytest
      biopython
 commands =
-    # Test if cy_module is built.
-    python3 -c 'import fastqsplitter.split_cy'
     # Create HTML coverage report for humans and xml coverage report for external services.
     coverage run --source=fastqsplitter -m py.test -v tests
     coverage html
@@ -23,15 +21,6 @@ commands =
     # Run flake8 on the .pyx files. Ignore SyntaxError (E999)
     flake8 --filename=*.pyx --ignore=E999 src/
     mypy src/fastqsplitter
-
-[testenv:nocy]
-setenv=NO_CYTHON=true
-whitelist_externals=bash
-commands =
-    # Test if split_cy is not built.
-    bash -c "python3 -c $'import sys\ntry:\n    import fastqsplitter.split_cy\n    sys.exit(1)\nexcept ImportError:\n    sys.exit(0)'"
-    # Test if usage displays a warning
-    bash -c 'fastqsplitter --help | grep WARNING'
 
 # Documentation should build on python version 3
 [testenv:docs]
